ReflectionDestroyerService as its own independent class
