+++ app/workers/delete_payment_setting_hierarchy_worker.rb
@@ -11,6 +11,6 @@ class DeletePaymentSettingHierarchyWorker < DeleteObjectHierarchyWorker
     return unless called_from_provider_hierarchy?
 
     reflection = BackgroundDeletion::Reflection.new([:buyer_accounts, { action: :destroy, class_name: 'Account' }])
-    ReflectionDestroyer.new(object.account, reflection, caller_worker_hierarchy).destroy_later
+    Deletion::ReflectionDestroyerService.call(object.account, reflection, caller_worker_hierarchy)
   end
 end
